Refactor EntityService to use a EntityRegistry helper class
