+++ Celbridge/CoreServices/Celbridge.BaseLibrary/Entities/PatchSummary.cs
@@ -1,7 +1,7 @@
 namespace Celbridge.Entities;
 
 /// <summary>
-/// Describes the changes applied by a patch to an entity data object.
-/// Includes a JSON Patch that can be used to reverse the changes.
+/// Describes the changes applied by a patch to an entity.
+/// Includes both the original JSON patch and a reverse patch that can be used to undo the changes.
 /// </summary>
 public record PatchSummary(string Patch, string ReversePatch, List<ComponentChangedMessage> ComponentChangeMessages);
